# Time and space optimization of address persistence