--- a/lib/core/test/unit/Cardano/Wallet/DB/Arbitrary.hs
+++ b/lib/core/test/unit/Cardano/Wallet/DB/Arbitrary.hs
@@ -103,6 +103,7 @@ import Cardano.Wallet.Primitive.Types
     , WalletName (..)
     , WalletPassphraseInfo (..)
     , flatSlot
+    , fromFlatSlot
     , isPending
     , rangeIsValid
     , slotSucc
@@ -148,7 +149,7 @@ import GHC.Generics
 import System.IO.Unsafe
     ( unsafePerformIO )
 import System.Random
-    ( mkStdGen )
+    ( mkStdGen, random, randoms )
 import Test.QuickCheck
     ( Arbitrary (..)
     , Gen
@@ -164,9 +165,11 @@ import Test.QuickCheck
     , generate
     , genericShrink
     , oneof
+    , resize
     , scale
     , shrinkIntegral
     , shrinkList
+    , sized
     , vector
     , vectorOf
     )
@@ -295,7 +298,7 @@ instance Arbitrary MockChain where
 instance GenState s => Arbitrary (InitialCheckpoint s) where
     shrink (InitialCheckpoint cp) = InitialCheckpoint <$> shrink cp
     arbitrary = do
-        cp <- arbitrary @(Wallet s)
+        cp <- resize 0 $ arbitrary @(Wallet s)
         pure $ InitialCheckpoint $ unsafeInitWallet
             (utxo cp)
             (block0 ^. #header)
@@ -342,9 +345,9 @@ instance Arbitrary PassphraseScheme where
 -------------------------------------------------------------------------------}
 
 instance Arbitrary BlockHeader where
-    arbitrary = do
-        sid@(SlotId (EpochNo ep) (SlotNo sl)) <- arbitrary
-        let h = fromIntegral sl + fromIntegral ep * arbitraryEpochLength
+    arbitrary = sized $ \n -> do
+        let h = fromIntegral n
+        let sid = fromFlatSlot (EpochLength arbitraryEpochLength) (fromIntegral h)
         blockH <- arbitrary
         pure $ BlockHeader sid (Quantity h) blockH (coerce blockH)
 
@@ -530,17 +533,31 @@ arbitraryRewardAccount =
                                  Random State
 -------------------------------------------------------------------------------}
 
+{- HLINT ignore "Use !!" -}
 instance Arbitrary (RndState 'Mainnet) where
     shrink (RndState k ix addrs pending g) =
         [ RndState k ix' addrs' pending' g
         | (ix', addrs', pending') <- shrink (ix, addrs, pending)
         ]
-    arbitrary = RndState
-        <$> pure (Passphrase "passphrase")
-        <*> pure minBound
-        <*> arbitrary
-        <*> (pure mempty) -- FIXME: see comment on 'Arbitrary Seq.PendingIxs'
-        <*> pure (mkStdGen 42)
+    -- Addresses are generate based on the size parameter, so that property
+    -- tests can actually expect addresses to not be totally random, but
+    -- instead, be an ever growing map, much more like they are actually
+    -- generated in practice.
+    arbitrary = sized $ \n -> do
+        let stdgen = mkStdGen 42
+        let newAddress (m, g0) = (Map.insert k v m, g2)
+              where
+                (acctIx, g1) = random g0
+                (addrIx, g2) = random g1
+                k = (Index acctIx, Index addrIx)
+                v = Address $ B8.pack $ take 32 $ randoms g2
+        let addrs = fst $ head $ drop n $ iterate newAddress (mempty, stdgen)
+        pure $ RndState
+            (Passphrase "passphrase")
+            minBound
+            addrs
+            mempty -- FIXME: see comment on 'Arbitrary Seq.PendingIxs'
+            stdgen
 
 instance Arbitrary (ByronKey 'RootK XPrv) where
     shrink _ = []