@KtorZ 2020-05-14
Note: this probably deserve some attention regarding pre-existing database. I assume we can simply get away by "nubbing" duplicates that would be selected :thinking: