@rvl 2020-05-20
The idea of writing diffs to the database sounds good. It's more suited to the wallet than "put" operations a la mongodb. I also had a branch which changed the return type of `applyBlocks` to be a checkpoint diff.