@KtorZ 2020-05-20
> I also had a branch which changed the return type of applyBlocks to be a checkpoint diff.

Yes, I recall you groomed on that a while ago about the UTxO. It sounded more difficult there though? Do you still have that somewhere?